+++ userfiles/modules/microweber/toolbar/editor_tools/module_settings/index.php
@@ -146,7 +146,7 @@
 
         if (typeof thismodal === 'undefined' && self !== parent && typeof this.name != 'undefined' && this.name != '') {
             var frame = parent.mw.$('#' + this.name)[0];
-            thismodal = parent.mw.tools.modal.get(mw.tools.firstParentWithClass(frame, 'mw_modal'));
+            thismodal = parent.mw.dialog.get(mw.tools.firstParentWithClass(frame, 'mw_modal'));
         }
 
 
@@ -178,21 +178,7 @@
                 thismodal.title(modal_title_str)
             }
 
-
-            $(window).load(function () {
-                if (thismodal.main && !!thismodal.main[0]) {
-
-                    if (autoSize) {
-                        parent.mw.tools.modal.resize("#" + thismodal.main[0].id, false, $('#settings-container').height() + 25, true);
-
-                        $(window.parent.window).bind("resize", function () {
-                            if (parent != null) {
-                                mw.parent().dialog.get("#" + thismodal.main[0].id).center()
-                            }
-                        });
-                    }
-                }
-            });
+            
         }
 
         $(window).load(function () {